Fix warnings.

+++ daemon/boardd/article.c
@@ -33,7 +33,7 @@ answer_file(struct evbuffer *buf, const char *path, struct stat *st,
     if (ck && cklen) {
 	char ckbuf[128];
 	snprintf(ckbuf, sizeof(ckbuf), "%d-%d", (int) st->st_dev, (int) st->st_ino);
-	if (cklen != strlen(ckbuf) || strncmp(ck, ckbuf, cklen) != 0) {
+	if ((size_t) cklen != strlen(ckbuf) || strncmp(ck, ckbuf, cklen) != 0) {
 	    ret = SELECT_ECHANGED;
 	    goto answer_file_errout;
 	}
